- refactored recipe out of the cli.py.
