+++ adr/recipes/backout_rate.py
@@ -10,7 +10,7 @@ Get information on the backout rate on autoland and mozilla-inbound over the giv
 """
 from __future__ import print_function, absolute_import
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
