+++ adr/recipes/code_coverage_by_suite.py
@@ -1,6 +1,6 @@
 from __future__ import print_function, absolute_import
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
