+++ adr/recipes/inspect.py
@@ -9,7 +9,7 @@ available tables. To see the columns in a table, run:
 """
 from __future__ import print_function, absolute_import
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
