+++ adr/recipes/try_efficiency.py
@@ -12,7 +12,7 @@ at preventing backouts. It is roughly:
 """
 from __future__ import print_function, absolute_import
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
