+++ adr/recipes/try_usage.py
@@ -13,7 +13,7 @@ from __future__ import print_function, absolute_import
 
 from collections import defaultdict, OrderedDict
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
