+++ adr/recipes/try_users.py
@@ -11,7 +11,7 @@ from __future__ import print_function, absolute_import
 
 from collections import defaultdict
 
-from ..cli import RecipeParser
+from ..recipe import RecipeParser
 from ..query import run_query
